requirements.txt
discord.py-self
flask